website fully indonesianed

+++ package/db/tmp/colte_db/sample_db.sql
@@ -147,7 +147,7 @@ CREATE TABLE `customers` (
   `imsi` varchar(16) NOT NULL,
   `raw_down` int(10) unsigned DEFAULT '0',
   `raw_up` int(10) unsigned DEFAULT '0',
-  `data_balance` int(10) DEFAULT '0',
+  `data_balance` bigint(15) DEFAULT '0',
   `balance` float(10,2) DEFAULT '0' COMMENT 'in USD for now',
   `bridged` tinyint(1) DEFAULT '0',
   `enabled` tinyint(1) DEFAULT '0',
